Factor out services into new Services component
